requirements.txt
# Flask & Web Framework
flask==3.0.0
flask-login==0.6.3
flask-mail==0.9.1
flask-wtf==1.2.1
flask-limiter==3.5.0
werkzeug==3.0.1

# Supabase Database
supabase==2.3.4
postgrest-py==0.16.3
gotrue==2.1.0
realtime==1.0.6
storage3==0.7.4
supafunc==0.5.1

# HTTP & API
requests==2.31.0
httpx==0.26.0
gunicorn==21.2.0

# Chinese Language Processing
pypinyin==0.50.0

# Text-to-Speech
gtts==2.5.1

# HTML Parsing & Web Scraping
beautifulsoup4==4.12.3
selenium==4.17.2

# Image Processing
pillow==10.2.0

# Database
psycopg2-binary==2.9.9

# Data Processing
python-dotenv==1.0.0

# Security
bcrypt==4.1.2
pyjwt==2.8.0
cryptography==42.0.0

# Utilities
python-slugify==8.0.3
cachetools==5.3.2

# Development (not needed for production)
pytest==7.4.4
pytest-flask==1.3.0
